+++ src/Farkle.Tools/Program.fs
@@ -20,29 +20,21 @@ type FarkleCLIExiter() =
 
 type Arguments =
     | Version
-    | ``Explain-composite-paths``
     | [<Inherit>] Json
     | [<Inherit; AltCommandLine("-v"); Unique>] Verbosity of LogEventLevel
-    | [<CliPrefix(CliPrefix.None)>] New of ParseResults<New.Arguments>
-#if TODO_PRECOMPILER
+    | [<CliPrefix(CliPrefix.None)>] Render of ParseResults<Render.Arguments>
+    | [<CliPrefix(CliPrefix.None); Hidden>] New of ParseResults<Render.Arguments>
     | [<CliPrefix(CliPrefix.None)>] List of ParseResults<List.Arguments>
-    | [<AltCommandLine("generate-predefined-sets"); Hidden>] GeneratePredefinedSets of ParseResults<GeneratePredefinedSets.Arguments>
-#endif
 with
     interface IArgParserTemplate with
         member x.Usage =
             match x with
             | Version -> "Display the program's version info."
-            | ``Explain-composite-paths`` -> "Display help about the syntax of composite paths."
             | Json -> "Encode output in JSON and print it in a single line in stdout. \
 No files will be created and only errors will be logged by default."
             | Verbosity _ -> "Set the verbosity of the tool's logs."
-            | New _ -> "Generate a skeleton program from a grammar file and a Scriban template."
-#if TODO_PRECOMPILER
+            | New _ | Render _ -> "Render an HTML page or a custom Scriban template from a grammar file."
             | List _ -> "List all precompiled grammars of an assembly."
-            | GeneratePredefinedSets _ -> "Generate an F# source file with GOLD Parser's predefined sets. \
-For internal use only."
-#endif
 
 [<EntryPoint>]
 let main argv =
@@ -70,20 +62,14 @@ let main argv =
             if results.Contains Version then
                 Log.Information("Version: {toolsVersion:l}", toolsVersion)
                 0
-            elif results.Contains ``Explain-composite-paths`` then
-                printfn "Composite paths specify both a file path and a precompiled grammar name. Their format is 'file_path::grammar_name'."
-                printfn "The file name can be an assembly, an EGT file, or a project. In the latter case, the project may be built if required."
-                printfn "If the file name is omitted (i.e. the path has the form of '::grammar_name'), Farkle will try to find a project file in the current directory."
-                printfn "If only the file name is specified (i.e. the path has the form 'file_path'), the file's only precompiled grammar will be chosen."
-                0
             else
                 match results.GetSubCommand() with
-                | New args -> New.run json args
-#if TODO_PRECOMPILER
-                | GeneratePredefinedSets args -> GeneratePredefinedSets.run args
+                | New args ->
+                    Log.Warning("The {Subcommand:l} subcommand is deprecated. Use {Replacement:l} instead.", "new", "render")
+                    Render.run json args
+                | Render args -> Render.run json args
                 | List args -> List.run json args
-#endif
-                | Version | Json | Verbosity _ | ``Explain-composite-paths`` -> Ok ()
+                | Version | Json | Verbosity _ -> Ok ()
                 |> function | Ok () -> 0 | Error () -> 1
         with
         | :? FileNotFoundException as e ->